+++ ios/flutter_qjs_next/Sources/flutter_qjs_next/FlutterQjsPlugin.m
@@ -1,15 +1,21 @@
-#import "FlutterQjsPlugin.h"
-#if __has_include(<flutter_qjs_next/flutter_qjs_next-Swift.h>)
-#import <flutter_qjs_next/flutter_qjs_next-Swift.h>
-#else
-// Support project import fallback if the generated compatibility header
-// is not copied when this plugin is created as a library.
-// https://forums.swift.org/t/swift-static-libraries-dont-copy-generated-objective-c-header/19816
-#import "flutter_qjs_next-Swift.h"
-#endif
+#import "include/FlutterQjsPlugin.h"
 
 @implementation FlutterQjsPlugin
-+ (void)registerWithRegistrar:(NSObject<FlutterPluginRegistrar>*)registrar {
-  [SwiftFlutterQjsPlugin registerWithRegistrar:registrar];
+
++ (void)registerWithRegistrar:(NSObject<FlutterPluginRegistrar> *)registrar {
+  FlutterMethodChannel *channel =
+      [FlutterMethodChannel methodChannelWithName:@"flutter_qjs_next"
+                                  binaryMessenger:[registrar messenger]];
+  FlutterQjsPlugin *instance = [[FlutterQjsPlugin alloc] init];
+  [registrar addMethodCallDelegate:instance channel:channel];
 }
+
+- (void)handleMethodCall:(FlutterMethodCall *)call result:(FlutterResult)result {
+  if ([@"getPlatformVersion" isEqualToString:call.method]) {
+    result([@"iOS " stringByAppendingString:[[UIDevice currentDevice] systemVersion]]);
+  } else {
+    result(FlutterMethodNotImplemented);
+  }
+}
+
 @end